renamed Agent nodes
