+++ nodes/agent/SetDefaultAgent.py
@@ -1,5 +1,5 @@
 from ...py.griptape_config import update_config_with_dict
-from .base_agent import BaseAgent
+from .BaseAgent import BaseAgent
 
 
 class gtUISetDefaultAgent(BaseAgent):
